Refactor radar location distance type and enhance radar screen UI
- Changed the type of `radarLocationDistanceMeters` from double to int in `MeshController` for consistency with usage in location settings.
- Updated the radar screen UI to conditionally display GPS reliability information, improving user feedback based on the fusion state.
- Adjusted layout to ensure proper spacing and visibility of radar confidence metrics based on GPS reliability.

+++ app/lib/controllers/mesh_controller.dart
@@ -39,7 +39,7 @@ class MeshController extends ChangeNotifier {
   /// espaciado para no agotar batería ni saturar la malla.
   static const Duration rescueInterval = Duration(minutes: 5);
   static const Duration radarLocationInterval = Duration(seconds: 20);
-  static const double radarLocationDistanceMeters = 15;
+  static const int radarLocationDistanceMeters = 15;
 
   final MeshPlatformService _platform;
   final MessageRepository _repository;
@@ -514,7 +514,7 @@ class MeshController extends ChangeNotifier {
           Geolocator.getPositionStream(
             locationSettings: const LocationSettings(
               accuracy: LocationAccuracy.high,
-              distanceFilter: radarLocationDistanceMeters.toInt(),
+          distanceFilter: radarLocationDistanceMeters,
             ),
           ).listen((position) {
             _latestRadarPosition = position;
